[project @ 2001-06-14 12:50:05 by simonpj] ---------------------- 	Installation packaging 	----------------------
GHC runs various system programs like
	cp, touch
	gcc, as, ld etc

On Windows we plan to deliver these programs along with GHC,
so we have to be careful about where to find them.

This commit isolates all these dependencies in a single module

	main/SysTools.lhs

Most of the #ifdefery for mingw has moved into this module.
There's some documentation in SysTools.lhs

Along the way I did lots of other cleanups.  In particular

  * There is no more 'globbing' needed when calling runSomething
  * All file removal goes via the standard Directory.removeFile
  * TmpFiles.hs has gone; absorbed into SysTools
  * Some DynFlag stuff has moved from DriverFlags to CmdLineOpts


Still to do:

  **	I'm a bit concerned that calling removeFile one at a time
	when deleting masses of split-object files is going to be
	rather slow

  **	GHC now expects to find split,mangle,unlit in
		libdir/extra-bin
	instead of just
		libdir

	So something needs to change in the Unix installation scripts

  **    The "ineffective C preprocessor" is a perversion and should die