[project @ 2001-06-26 08:13:04 by reid] NOTE: new configure.in - remember to autoconf and reconfig
I'm moving X.gc and Xlib.gc from hslibs/graphics/lib/x11 to
hslibs/xlib and integrating them into the hslibs maketree.  This lib
now lives at the same place in the hierarchy as win32 - probably not
perfect for the new library story but good enough for the old hslibs I
think.

I'm not converting them from GreenCard to hsc2hs but I think that
would be a good medium term goal.  I'd like to do it myself because
I'd like to try using hsc2hs but I can't imagine when I'll get the
time so if someone feels like doing it, go right ahead.  (The HGL
(hslibs/graphics/lib/x11) is probably the only code that depends on
the Xlib interface - so mild changes to the Xlib API would be fairly
easy to fix up.)

It all seems to build and compile (not sure about linking yet) but,
for the life of me, I can't figure out what part of the makefile calls
ghc-pkg -a so that isn't quite working yet.  (I _think_ the makefile
does this - but could be wrong.)

If someone could either point me at the relevant part of the makefile
docs or just fix it for me, I could move onto getting the graphics lib
integrated into the hslibs maketree - which would leave me with just
the Hugs part of the HGL distribution to fix before the next Hugs
release.  (If I was a good Haggis, I'd beat up on Hugs' ffi too - but
I'm a very overworked Haggis at the moment so I'm not sure I'll get
that far.)