[project @ 2001-06-27 16:38:17 by simonmar] When we're in --interactive or --make mode, we don't even *look* for interface files in the home package.
This means that cd'ing into fptools/ghc/lib/std and starting up GHCi
Just Works, which is a good thing.  It also subsumes the previous hack
about checking whether we're renaming a command-line expression before
allowing a home interface to be loaded.

The downside is that if you try to use a qualified name for a home
module that's not loaded, you'll get a slightly less informative error
message: "interface file not found" rather than "module not loaded",
but this could be improved.