[project @ 2001-07-13 13:29:56 by simonpj] ------------------------------------ 	Tidy up the "syntax rebinding" story 	------------------------------------
I found a bug in the code that dealt with re-binding implicit
numerical syntax:
	literals 	(fromInteger/fromRational)
	negation	(negate)
	n+k patterns	(minus)

This is triggered by the -fno-implicit-prelude flag, and it
used to be handled via the PrelNames.SyntaxMap.

But I found a nicer way to do it that involves much less code,
and doesn't have the bug.  The explanation is with
	RnEnv.lookupSyntaxName