[project @ 2001-07-17 15:28:30 by simonpj] -------------------------------- 	First cut at the demand analyser 	--------------------------------
This demand analyser is intended to replace the strictness/absence
analyser, and the CPR analyser.

This commit adds it to the compiler, but in an entirely non-invasive
way.

	If you build the compiler without -DDEBUG,
	you won't get it at all.

	If you build the compiler with -DDEBUG,
	you'll get the demand analyser, but the existing
	strictness analyser etc are still there.  All the
	demand analyser does is to compare its output with
	the existing stuff and report differences.

There's no cross-module stuff for demand info yet.

The strictness/demand info is put the IdInfo as
	newStrictnessInfo
	newDemandInfo

Eventually we'll remove the old ones.

Simon