[project @ 2001-07-23 10:54:46 by simonpj] --------------------------------- 	Switch to the new demand analyser 	---------------------------------
This commit makes the new demand analyser the main beast,
with the old strictness analyser as a backup.  When
DEBUG is on, the old strictness analyser is run too, and the
results compared.

WARNING: this isn't thorougly tested yet, so expect glitches.
Delay updating for a few days if the HEAD is mission critical
for you.

But do try it out.  I'm away for 2.5 weeks from Thursday, so
it would be good to shake out any glaring bugs before then.