[project @ 2001-07-23 22:52:33 by ken] Improved how we build library archives:
On Digital UNIX (or where otherwise available), use the "Z" option to
compress archives as we build them.

On Digital UNIX 5 (or where otherwise available), use the "-input" option,
which reads from a specified file the list of objects to archive.  This is
significantly faster than piping the list of objects to xargs when the
number of objects to archive is large.