[project @ 2001-07-25 09:19:41 by simonpj] --------------------------------- 	Another long-standing infelicity! 	---------------------------------
CoreTidy was throwing away demand info on let-binders.
This meant that CorePrep would build a let (thunk) instead of
a case, even if the Id is sure to be demanded.

Easily fixed by making CoreTidy retain the demand info.

This demand-analysis stuff is having the excellent side effect
of flushing out performance bugs!