[project @ 2001-07-31 10:06:25 by sewardj] Disable all strictness analysis in the HEAD compiler until such time as we can figure out why the new analyser is not working correctly. I realise this is a drastic measure, but all previous attempts to either fix or work around the problem have failed.
If you are doing work which requires maxed-out performance, you may
want to revert the effects of this commit in your source tree.  The
compiler only generates wrong code under pretty obscure circumstances
(when compiling itself).  All other programs appear to work correctly
with the new analyser.

We hope this situation will not last for long.