[project @ 2001-08-02 11:20:50 by simonmar] Add a POSIX regular expression binding as Text/Regex/Posix.  POSIX regexps are provided virtually everywhere (except Windows, but there's always pcre), and this means we won't have to ship the a copy of GNU regex.c, which is GPL'ed.
Compared to hslibs/text/Regex.lhs, this one is much shorter (150 lines
vs. 350), more portable (uses the FFI libs and hsc2hs vs. casm), and
easier to use (it returns the right number of subexpressions from the
regex, rather than 32 all the time).

Text.Regex is a re-implementation of the old
hslibs/text/RegexString.lhs, written to use Text.Regex.Posix.  The
syntax of regexps is probably not identical, but it's certainly
similar enough that the URI library works unchanged with the new
implementation.