[project @ 2001-08-20 14:10:02 by simonmar] Remove the in-range assertions on mkMachInt/mkMachWord.  They clearly aren't true, because there's nothing stopping you from writing an out-of-range Int# literal (although that's the only way I can see for these to arise).
The wider issue is what should be done about out-of-range Int#
literals; I vaguely remember that at some point we disallowed them,
but I can't find anything in the logs.  The case which triggered the
assertion, namely "intToWord# 0xffff0000" would appear to be a
legitimate use for an out-of-range Int# literal though, given that
you can't write Word# literals directly.