[project @ 2001-08-21 09:25:48 by simonmar] Accept The Truth and disallow loading any home modules which don't have a source file.
The reason is that the compilation manager can't figure out
dependencies unless it has the source file to browse through -
previously there was a half-hearted attempt to let you get away with
just an object + interface file, but it didn't work in general.
Better just to disallow it altogether.

Note that this applies to --make too.