[project @ 2001-08-23 14:38:31 by rrt] The new InstallShield for Windows Installer installer. I won't nuke the old IS6 one for now, but expect it to go after the GHC 5.02 release.
The new installer uses the standard Windows Installer system, so should be
easier to migrate away from InstallShield to a more automated system in
future, though that will require the setup to be exported (the .ism format
is proprietary to InstallShield).

Also, the entire setup lives in one file rather than a zillion scattered
over a hierarchy with lots of empty directories and space-including
directory names, so it's rather easier to check out!