[project @ 2001-09-07 12:43:28 by simonpj] ----------------------------------- 	Pin on accurate strictness info for 	record and dictionary selectors 	-----------------------------------         [part of 3 related commits]
This fixes a long-standing infelicity.  Sometimes selectors aren't
inlined until after strictness analysis, so if we don't have decent
strictness info on them we get bad strictness results.

For record selectors, the unboxing-strict-fields stuff makes it hard
to figurwe out the correct strictness, so we just invoke the demand
analyser to work it out.