[project @ 2001-09-10 10:07:21 by rje] Fixed a bug in TICKY_TICKY profiling.
Previously, the TICK_ENT_DIRECT event was logged before the heap/stack
check was done. As a result, if the check failed, the TICK_ENT_DIRECT
event would be logged a second time, causing TICKY_TICKY to give innacurate
numbers.

This patch shouldn't have any affect on non-ticky compilation.

Also changed the modified bit of code to use "do" notation, and so look a bit
neater.